⬆️ Collection Products
Type(s): UPDATE

Issue(s):
- JIRA: https://blabol.atlassian.net/browse/MSBE-548

+++ app/Services/Collection/ImageUpdater.php
@@ -32,8 +32,7 @@ class ImageUpdater
             $fileName = $this->collection_repo->getDeletionFileNameCollectionImageFromCDN($partner_id, $collection_id, $column_name);
             $mainFileName = $this->getMainFileName($fileName);
 
-            if($mainFileName != 'default.jpg')
-                $this->deleteFileFromCDN(substr($fileName, strlen(config('s3.url'))));
+            if($mainFileName != 'default.jpg') $this->deleteFileFromCDN(substr($fileName, strlen(config('s3.url'))));
         }
     }
